// src/modules/forms/application/dto/forms.dto.ts
export interface ContactFormEmailData {
    name: string;
    job: string;
    email: string;
    phone: string;
    message: string;
}

export interface DemoRequestEmailData {
    name: string;
    company: string;
    email: string;
    country: string;
    industry: string;
    robotsOfInterest: string[];
}